Fix LibGDX Atlas parsing
(1) Texture Packer (not the one provided by LibGDX) isn't inserting
an initial line terminator. The initial line terminator is no longer
required.

(2) The scanner has now a 'generous' mode in which it will
accept anything as an identifier that's no white space and no
line terminator. This is useful when we know we're expecting an
identifier since it allows us to support identifiers starting with
digits and distinguish them from actual integers